Вынесла функцию цитирования в модуль
